colocando tests na pasta correta, estavam perdidos em outras pastas
